Move activity rights creation to webacl middleware
